tidy

+++ gen-imperium.cmd
@@ -1,2 +1,3 @@
-rd /s /q "./_Imperium/_Generated"
+if exist ".\_Imperium\_Generated\"  rd /s /q "./_Imperium/_Generated"
+timeout /t 5
 dotnet-codegen ./Specs/imperium.yaml ./Templates ./_Imperium
